Fix setupaliases.sh

+++ setupaliases.sh
@@ -122,15 +122,20 @@ fi
 # #############################################################################
 # Docker
 
-alias dc='docker-compose'
-alias dce='docker-compose exec'
-alias dk='docker'
-alias dke='docker exec'
-alias dkipaddr="docker inspect --format '{{ .NetworkSettings.IPAddress }}'"
-​alias dkl='docker ps -l -q'
-​alias dkps="docker ps"
-alias dkrm="docker rm"
-alias dkrmi="docker rmi"
+if which docker-compose >/dev/null 2>&1 ; then
+  alias dc='docker-compose'
+  alias dce='docker-compose exec'
+fi
+
+if which docker >/dev/null 2>&1 ; then
+  alias dk='docker'
+  alias dke='docker exec'
+  alias dkipaddr="docker inspect --format '{{ .NetworkSettings.IPAddress }}'"
+  alias dkl='docker ps -l -q'
+  alias dkps="docker ps"
+  alias dkrm="docker rm"
+  alias dkrmi="docker rmi"
+fi
 
 # #############################################################################
 # Git
